refactor(projects): правки после ревью
Jira issue: VEGA-551
