+++ src/pages/projects/ProjectsTable/EditedAt/index.ts
@@ -1,0 +1,1 @@
+export { EditedAt } from './EditedAt';
